--- src/diffie_hellman.cc
@@ -1,25 +1,0 @@
-//
-// Created by wtchr on 8/5/2024.
-//
-
-#include "diffie_hellman.h"
-
-#include "mpz.h"
-
-const auto p_value = mpz_class{
-        "0xFFFFFFFFFFFFFFFFADF85458A2BB4A9AAFDC5620273D3CF1D8B9C583CE2D3695A9E13641146433FBCC939DCE249B3EF97D2FE363630C"
-        "75D8F681B202AEC4617AD3DF1ED5D5FD65612433F51F5F066ED0856365553DED1AF3B557135E7F57C935984F0C70E0E68B77E2A689DAF3"
-        "EFE8721DF158A136ADE73530ACCA4F483A797ABC0AB182B324FB61D108A94BB2C8E3FBB96ADAB760D7F4681D4F42A3DE394DF4AE56EDE7"
-        "6372BB190B07A7C8EE0A6D709E02FCE1CDF7E2ECC03404CD28342F619172FE9CE98583FF8E4F1232EEF28183C3FE3B1B4C6FAD733BB5FC"
-        "BC2EC22005C58EF1837D1683B2C6F34A26C1B2EFFA886B423861285C97FFFFFFFFFFFFFFFF"};
-
-diffie_hellman::diffie_hellman()
-    : p{p_value}
-    , g{2}
-    , x{random_prime(255)}
-    , y{powm(g, x, p)} {}
-
-mpz_class diffie_hellman::set_peer_public_key(const mpz_class &pub_key) {
-    this->K = powm(pub_key, x, p);
-    return K;
-}
