+++ src/diffie_hellman.cpp
@@ -1,12 +1,32 @@
 //
-// Created by wtchr on 8/8/2024.
+// Created by wtchr on 8/5/2024.
 //
 
-#include "ecdsa.h"
+#include "diffie_hellman.h"
 
 #include <cassert>
 #include <vector>
 #include "mpz.h"
+
+// diffie_hellman
+
+const auto p_value = mpz_class{
+        "0xFFFFFFFFFFFFFFFFADF85458A2BB4A9AAFDC5620273D3CF1D8B9C583CE2D3695A9E13641146433FBCC939DCE249B3EF97D2FE363630C"
+        "75D8F681B202AEC4617AD3DF1ED5D5FD65612433F51F5F066ED0856365553DED1AF3B557135E7F57C935984F0C70E0E68B77E2A689DAF3"
+        "EFE8721DF158A136ADE73530ACCA4F483A797ABC0AB182B324FB61D108A94BB2C8E3FBB96ADAB760D7F4681D4F42A3DE394DF4AE56EDE7"
+        "6372BB190B07A7C8EE0A6D709E02FCE1CDF7E2ECC03404CD28342F619172FE9CE98583FF8E4F1232EEF28183C3FE3B1B4C6FAD733BB5FC"
+        "BC2EC22005C58EF1837D1683B2C6F34A26C1B2EFFA886B423861285C97FFFFFFFFFFFFFFFF"};
+
+diffie_hellman::diffie_hellman()
+    : p{p_value}
+    , g{2}
+    , x{random_prime(255)}
+    , y{powm(g, x, p)} {}
+
+mpz_class diffie_hellman::set_peer_public_key(const mpz_class &pub_key) {
+    this->K = powm(pub_key, x, p);
+    return K;
+}
 
 // ec_field
 
@@ -86,58 +106,4 @@ ec_point operator*(const mpz_class &l, const ec_point &p) {
 std::ostream &operator<<(std::ostream &os, const ec_point &r) {
     os << "(" << r.x << ", " << r.y << ")";
     return os;
-}
-
-// ecdsa
-
-ecdsa_class::ecdsa_class(const ec_point &G, mpz_class n)
-    : ec_point{G} {
-    this->n = n;
-    this->n_bit = mpz_sizeinbase(n.get_mpz_t(), 2);
-}
-
-mpz_class ecdsa_class::mod_inv(const mpz_class &z) const {
-    mpz_class r;
-    mpz_invert(r.get_mpz_t(), z.get_mpz_t(), n.get_mpz_t());
-    return r;
-}
-
-std::pair<mpz_class, mpz_class> ecdsa_class::sign(const mpz_class &m, const mpz_class &d) const {
-    // Discard last bits if m is too big
-    const size_t m_bit = mpz_sizeinbase(m.get_mpz_t(), 2);
-    const mpz_class z = m >> std::max(static_cast<int>(m_bit - n_bit), 0);
-
-    mpz_class k, s, r;
-    ec_point P = *this;
-    do {
-        do {
-            k = random_prime(31);
-            P = k * *this; // k * G
-            r = P.x % n;
-        } while (r == 0);
-        s = mod_inv(k) * (z + r * d) % n;
-    } while (s == 0);
-    return {r, s};
-}
-
-bool ecdsa_class::verify(const mpz_class &m, const std::pair<mpz_class, mpz_class> &sig, const ec_point &Q) const {
-    auto [r, s] = sig;
-    if (r < 1 || r >= n)
-        return false;
-    if (s < 1 || s >= n)
-        return false;
-
-    // Discard last bits if m is too big
-    const size_t m_bit = mpz_sizeinbase(m.get_mpz_t(), 2);
-    const mpz_class z = m >> std::max(static_cast<int>(m_bit - n_bit), 0);
-
-    const mpz_class inv_s = mod_inv(s);
-    const mpz_class u = z * inv_s % n;
-    const mpz_class v = r * inv_s % n;
-    const ec_point P = u * *this + v * Q;
-    if (P.is_identity())
-        return false;
-    if ((P.x - r) % n == 0)
-        return true;
-    return false;
 }
